requirements.txt
# Maily BASE Requirements
# Generated on 2025-02-25 20:20:49

# Core Backend Framework
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0.post2

# Database
sqlalchemy==2.0.12
alembic==1.10.4
psycopg2-binary==2.9.6
asyncpg==0.28.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Caching & Messaging
redis==4.5.5
aioredis==2.0.1
pika==1.3.2

# HTTP & API
requests==2.30.0
aiohttp==3.8.5

# Testing
httpx==0.24.0
factory-boy==3.3.0

# Monitoring
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0

# Utilities
loguru==0.7.0
tenacity==8.2.2
python-dateutil==2.8.2

# Code Quality

# Vault integration
hvac==1.2.1
